+++ lib/darujme_cz/address.rb
@@ -10,6 +10,11 @@ module DarujmeCz
       "#{@source["street"]}\n#{@source["postCode"]} #{@source["city"]}\n#{@source["country"]}"
     end
 
+    # alias
+    def postal_code
+      post_code
+    end
+
     %w[street postCode city country].each do |m|
       define_method m.underscore do
         @source[m]