+++ lib/darujme_cz/donor.rb
@@ -1,7 +1,7 @@
 module DarujmeCz
   class Donor
 
-    delegate :city, :street, :postal_code, :country, to: :address
+    delegate :city, :street, :post_code, :postal_code, :country, to: :address
     # @param [Hash] attributes
     def initialize(attributes)
       @source = attributes